% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidHymo.R
\name{braidHymo}
\alias{braidHymo}
\title{Calculate morphological indexes multiple rivers or reaches}
\usage{
braidHymo(rivers, rel_path = "")
}
\arguments{
\item{rivers}{a table containing river-related details.}
}
\value{
a table of morphological indexes
}
\description{
Calculate morphological indexes multiple rivers or reaches
}
\examples{
rivers=readr::read_csv2("data-raw/rivers.csv") \%>\%
   mutate(filepath=paste0("data-raw/",filepath))
braidHymo(rivers)
}
