% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidHymo_one.R
\name{braidHymo_one}
\alias{braidHymo_one}
\title{Calculate morphological indexes for one river or reach}
\usage{
braidHymo_one(rivdata, area, points_space)
}
\arguments{
\item{rivdata}{river data}

\item{area}{total area}

\item{points_space}{space between points}
}
\value{
a 1-row table of morphological indexes
}
\description{
Calculate morphological indexes for one river or reach
}
\examples{
data_Drac=braidHymo_read("data-raw/Drac_Chabottes_2018.txt")
result_Drac=braidHymo_one(data_Drac,
                          area=253,
                          points_space=1)
result_Drac
}
