% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidHymo_plot.R
\name{braidHymo_plot}
\alias{braidHymo_plot}
\title{Plot index (either W`*` or BRI`*`)}
\usage{
braidHymo_plot(
  tib_morpho_index,
  index = "BRI*",
  position,
  color,
  threshold = "default"
)
}
\arguments{
\item{tib_morpho_index}{the table with morphological indexes values, resulting from a call to braidHymo().}

\item{index}{the index used (can be either "W*" or "BRI*". Defaults to "BRI*").}

\item{position}{the variable according to which estimates should be placed on graphic.}

\item{color}{the variable according to which estimates should be colored.}

\item{threshold}{the threshold for a section to be considered as braiding (defaults to "default", i.e. 0.004 if the index considered is BRI or 10 if the index considered is W*).}
}
\value{
plot of index
}
\description{
Plot index (either W`*` or BRI`*`)
}
\examples{
rivers=readr::read_csv("data-raw/rivers.csv") \%>\%
   dplyr::mutate(filepath=paste0("data-raw/",filepath)) \%>\%
   dplyr::mutate(year=as.factor(year))
result=braidHymo(rivers)
braidHymo_plot(result,index="BRI*", position=year, color=river)
braidHymo_plot(result,index="W*", position=year, color=river)
}
