% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braidHymo_read.R
\name{braidHymo_read}
\alias{braidHymo_read}
\title{Read rivdata file}
\usage{
braidHymo_read(
  file,
  ID_XS = ID,
  Z = RASTERVALU,
  NA.value = -9999,
  sep = ";",
  row.names = 1
)
}
\arguments{
\item{ID_XS}{column corresponding to ID of cross-section in the file (defaults to ID_XS)}

\item{Z}{column corresponding the altitude in the file (defaults to Z)}

\item{NA.value}{value of missing values in the file (defaults to -9999)}

\item{filepath}{path to rivdata file}
}
\value{

}
\description{
Read rivdata file
}
\examples{
braidHymo_read("data-raw/Drac_Chabottes_2018.txt")
braidHymo_read("data-raw/Durance_Brillanne_2017.txt")
}
